SQL Parsing: Create, Insert, Drop and select and InMemory Catalog - In Memory Catalog. - Multy catalog support. - In Memory stream fetch

+++ polarspark/sql/_internal/parser/ddl.py
@@ -1,8 +1,8 @@
-from typing import  TYPE_CHECKING
+from typing import TYPE_CHECKING
 from shutil import rmtree
 import sqlglot.expressions as expr
 
-from polarspark.sql._internal.parser.models import CreateTable, InsertInto, DropTable
+from polarspark.sql._internal.parser.models import SourceTable, InsertInto, DropTable
 from polarspark.sql.types import _parse_datatype_string
 from polarspark.sql._internal.parser.parser import parse_sql
 from polarspark.sql._internal.parser import ast
@@ -10,18 +10,13 @@ from polarspark.sql._internal.parser import ast
 if TYPE_CHECKING:
     from polarspark.sql import SparkSession
 
-AST_TYPE_MAP = {
-    "TEXT": "STRING"
-}
+AST_TYPE_MAP = {"TEXT": "STRING"}
 
 
-def execute_create_table(spark: "SparkSession", ctx: CreateTable):
+def execute_create_table(spark: "SparkSession", ctx: SourceTable):
     table_name = ctx.name
-    schema = ["{} {}".format(col, AST_TYPE_MAP.get(ty, ty))
-              for col, ty
-              in ctx.columns]
+    schema = ["{} {}".format(col, AST_TYPE_MAP.get(ty, ty)) for col, ty in ctx.columns]
     schema = ", ".join(schema)
-    print(f"Schema is: {schema} {_parse_datatype_string(schema)}")
     spark.catalog.createTable(
         tableName=table_name,
         path=ctx.location,
@@ -31,25 +26,25 @@ def execute_create_table(spark: "SparkSession", ctx: CreateTable):
 
 
 def execute_insert_into(spark: "SparkSession", ctx: InsertInto):
-    tbl = spark.catalog._cat.get_ts(ctx.table).tables[ctx.table]
-    print(f"Table {ctx.table} is: {tbl}")
-    print(f"Ctx {ctx}")
-    schema = ["{} {}".format(col, AST_TYPE_MAP.get(ty, ty))
-              for col, ty
-              in tbl.columns if col in ctx.columns]
+    tbl = spark.catalog._cat.get_ts(ctx.table).tables[ctx.table]  # noqa
+    schema = [
+        "{} {}".format(col, AST_TYPE_MAP.get(ty, ty))
+        for col, ty in tbl.columns
+        if col in ctx.columns
+    ]
 
-    print(ctx)
     df = spark.createDataFrame(ctx.values or [], schema=schema)
     df.show()
     df.write.mode("append").format(tbl.format).save(tbl.location)
 
 
 def execute_drop_table(spark: "SparkSession", ctx: DropTable):
-    ts = spark.catalog._cat.get_ts(ctx.table)
+    ts = spark.catalog._cat.get_ts(ctx.table)  # noqa
     if ctx.table in ts.tables:
         tbl = ts.tables.pop(ctx.table)
         ts.pl_ctx.unregister(ctx.table)
-        rmtree(tbl.location)
+        if tbl.location:
+            rmtree(tbl.location)
 
 
 def execute_sql(spark: "SparkSession", sql: str):
